Add `Item` which has more fidelity about what value is changing.
It knows the key-value pair for slices of a mapping.
And for ArrayElement it know the index location.

This should make it easier to calculate sizes:

┌─> person:
│     name: Foo
└─>   surname: Bar

If we only get the height of the `value`, it would be 2.
But the key itself is on a different line, so its actually 3.